Finished updating all figures to SVG
Ultimately, converting text to outlines made for more consistent output, independent of which fonts are installed on a given machine.

Note:  .html files were added to the .gitignore file in a previous commit.  That makes sense for source control but not for our ultimate goal of using the HTML output in github.io.