Updated tables in Chapters 2-3
Note that the width was handled with the %autowidth feature.  This feature may not be available or appropriate for all outputs.  For example, how does it work for a book?

Another simplification was using the l (lower case L) formatting marker for columns to mark them as literal (and thus needing no backtick markup for individual cells).