Updated exercises to auto-number
Changed exercises to auto-number (using . instead of an explicit list numbering).

Doing so requires [start=x] where x is the starting number for a section of listing.  Setting the starting point is only necessary after a Programming Practice or Experiments marker.

The good news is that Conceptual Problems, Programming Practice, and Experiments are all now correctly indented.  Renumbering exercises will also be easier.

The bad news is that references to exercises still must be hand-numbered.