Fixed chapter-based xrefs
Now, programs, tables, figures, examples are all relative to the current chapter instead of global.

I.e., a figure will be numbered Figure 3.4 if it's the fourth figure in Chapter 3, instead of Figure 27 (or similar).

To access this functionality, one must run update-html.rb either directly with ruby or by using the update-html script.

I think I can extend this magic to exercises as well, allowing the exercises to be numbered automatically instead of by hand, but it won't be as easy.