Updated exercises to autonumber
Ruby code in autoxref-treeprocessor.rb now autonumbers exercises.  It was necessary to remove all of the explicit numbering in each chapter.  In the future, exercises can be added and removed with minimal difficulty since all numbering and cross-referencing is handled by the system.

Note that a little Ruby magic makes the lists of exercises automatically re-start numbering at the correct points without using [start=13] or similar syntax, after text like Programming Practice or Experiments.  The magic applies to top-level lists in a top-level section called Exercises.