Added auto-xref extension to PDF commands
Without auto-xref, the PDF numbering will not match the HTML numbering.  Worse, the exercises will have ugly links and restart numbering under each subheading (such as Programming Practice).

I updated `update-pdf` to use `autoxref-treeprocessor.rb`.  I also added an `update.pdf.rb` to allow direct Ruby interaction with asciidoctor-pdf, if that's desirable. Both seem to work.

Note: The syntax highlighting is different for the PDF version.  It balked when I tried to include the custom theme in `default.rb`.  It seems that asciidoctor-pdf uses Rouge but a *different* version of Rouge that it includes itself.