Added callouts to programs included in several segments
Until recently, it was very difficult to add callouts to a program that was included in several segments.

Numbering the callouts inside the program sequentially doesn't work.  If a program has callouts <1> and <2> in its first segment and <3> and <4> in the second, Asciidoctor complains when the <3> and <4> are listed to describe the second section, since each block of callout numbers are supposed to start at <1>.

The solution adopted by this commit is to use callout autonumbering (using <.> for all callouts).  The benefits of callout autonumbering are that they don't' require the author to keep track of specific numbers and each segment of code automatically starts at <1>.  This solution works even if included code segments overlap.

In my opinion, however, it would be nice if callout numbering could start at particular points, to show progression through a program.  After a comment thread with Dan Allen, I intend to make an issue request for this feature in Asciidoctor.