Spell-checking and compilation of example programs
This was a final run of polish for the latest version of the book.

At this point, the HTML version is in pretty good shape.
The PDF version no longer has issues with math, but the figures are strangely sized and not centered.  Sometimes the PDF figures are missing colors or graphics.

I'm using an older version of asciidoctor-pdf.  The newer one will fix some of these problems.  For others, we might need to use PDF versions of the figures instead of SVG.